fixed length calculation

+++ src/main/java/io/github/sk8erboi17/transformers/Markers.java
@@ -3,8 +3,8 @@ package io.github.sk8erboi17.transformers;
 public final class Markers {
 
     public static final byte START_MARKER = 0x01;
-    public static final int DATA_TYPE_SIZE = 1;
-    public static final int START_MARKER_SIZE = 1;
+    public static final byte DATA_TYPE_SIZE = 1;
+    public static final byte START_MARKER_SIZE = 1;
 
     private Markers() {
     }
